requirements.txt
# Core dependencies for weather intelligence platform
azure-eventhub>=5.11.4
azure-kusto-data>=4.3.1
azure-kusto-ingest>=4.3.1
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-monitor-opentelemetry>=1.2.0

# Data processing and analytics
databricks-connect>=13.3.0
pyspark>=3.5.0
pandas>=2.1.4
numpy>=1.24.3
scipy>=1.11.4

# Machine learning
scikit-learn>=1.3.2
xgboost>=2.0.3
statsmodels>=0.14.1
lightgbm>=4.3.0
optuna>=3.5.0

# Time series analysis
prophet>=1.1.5
pmdarima>=2.0.4
arch>=6.2.0

# Data validation and quality
great-expectations>=0.18.8
pandera>=0.17.2

# API and web framework
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.2
httpx>=0.26.0

# Configuration management
pyyaml>=6.0.1
python-dotenv>=1.0.0
hydra-core>=1.3.2

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=23.2.0
sentry-sdk>=1.39.2

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.2
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development tools
black>=23.12.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0

# Jupyter and notebooks
jupyter>=1.0.0
jupyterlab>=4.0.9
ipykernel>=6.28.0

# Visualization (for development)
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0

# Performance and profiling
psutil>=5.9.7
memory-profiler>=0.61.0

# Utilities
tqdm>=4.66.1
click>=8.1.7
rich>=13.7.0